+++ pkg/drivers/commonconfig_test.go
@@ -5,33 +5,17 @@ import (
 	"testing"
 
 	csmv1 "github.com/dell/csm-operator/api/v1"
-	"github.com/dell/csm-operator/pkg/utils"
-	"github.com/dell/csm-operator/tests/shared"
-	"github.com/dell/csm-operator/tests/shared/crclient"
+	//"github.com/dell/csm-operator/tests/shared/crclient"
 	"github.com/stretchr/testify/assert"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	//"k8s.io/apimachinery/pkg/runtime"
+	//"sigs.k8s.io/controller-runtime/pkg/client"
+	//"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
 
 var (
 	//
-	csm                                       = csmWithTolerations()
 	fakeDriver               csmv1.DriverType = "fakeDriver"
 	badDriver                csmv1.DriverType = "badDriver"
-	powerScaleCSM                             = csmForPowerScale()
-	powerScaleCSMBadSkipCert                  = csmForPowerScaleBadSkipCert()
-	powerScaleCSMBadCertCnt                   = csmForPowerScaleBadCertCnt()
-	powerScaleCSMBadVersion                   = csmForPowerScaleBadVersion()
-	objects                                   = map[shared.StorageKey]runtime.Object{}
-	powerScaleClient                          = crclient.NewFakeClientNoInjector(objects)
-	powerScaleSecret                          = shared.MakeSecret("csm-creds", "driver-test", shared.ConfigVersion)
-
-	// where to find all the yaml files
-	config = utils.OperatorConfig{
-		ConfigDirectory: "../../tests/config",
-	}
 
 	tests = []struct {
 		// every single unit test name
@@ -49,87 +33,7 @@ var (
 		{"file does not exist", csm, fakeDriver, "NonExist.yaml", "no such file or directory"},
 		{"config file is invalid", csm, badDriver, "bad.yaml", "unmarshal"},
 	}
-
-	powerScaleTests = []struct {
-		// every single unit test name
-		name string
-		// csm object
-		csm csmv1.ContainerStorageModule
-		// client
-		ct  client.Client
-		sec *corev1.Secret
-		// expected error
-		expectedErr string
-	}{
-		{"happy path", powerScaleCSM, powerScaleClient, powerScaleSecret, ""},
-		{"invalid value for skip cert validation", powerScaleCSMBadSkipCert, powerScaleClient, powerScaleSecret, "is an invalid value for X_CSI_ISI_SKIP_CERTIFICATE_VALIDATION"},
-		{"invalid value for cert secret cnt", powerScaleCSMBadCertCnt, powerScaleClient, powerScaleSecret, "is an invalid value for CERT_SECRET_COUNT"},
-	}
-
-	preCheckPowerScaleTest = []struct {
-		// every single unit test name
-		name string
-		// csm object
-		csm csmv1.ContainerStorageModule
-		// client
-		ct client.Client
-		// secret
-		sec *corev1.Secret
-		// expected error
-		expectedErr string
-	}{
-		{"missing secret", powerScaleCSM, powerScaleClient, powerScaleSecret, "failed to find secret"},
-		{"bad version", powerScaleCSMBadVersion, powerScaleClient, powerScaleSecret, "not supported"},
-	}
-
-	opts = zap.Options{
-		Development: true,
-	}
-
-	// logger = zap.New(zap.UseFlagOptions(&opts)).WithName("pkg/drivers").WithName("unit-test")
-
-	trueBool  bool = true
-	falseBool bool = false
 )
-
-func TestGetApplyCertVolume(t *testing.T) {
-	for _, tt := range powerScaleTests {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := getApplyCertVolume(tt.csm)
-			if tt.expectedErr == "" {
-				assert.Nil(t, err)
-			} else {
-				assert.Containsf(t, err.Error(), tt.expectedErr, "expected error containing %q, got %s", tt.expectedErr, err)
-			}
-		})
-	}
-}
-
-func TestPrecheckPowerScale(t *testing.T) {
-	ctx := context.Background()
-	for _, tt := range preCheckPowerScaleTest {
-		t.Run(tt.name, func(t *testing.T) {
-			err := PrecheckPowerScale(ctx, &tt.csm, config, tt.ct)
-			if tt.expectedErr == "" {
-				assert.Nil(t, err)
-			} else {
-				assert.Containsf(t, err.Error(), tt.expectedErr, "expected error containing %q, got %s", tt.expectedErr, err)
-			}
-		})
-	}
-
-	for _, tt := range powerScaleTests {
-		tt.ct.Create(ctx, tt.sec)
-		t.Run(tt.name, func(t *testing.T) {
-			err := PrecheckPowerScale(ctx, &tt.csm, config, tt.ct)
-			if tt.expectedErr == "" {
-				assert.Nil(t, err)
-			} else {
-				assert.Containsf(t, err.Error(), tt.expectedErr, "expected error containing %q, got %s", tt.expectedErr, err)
-			}
-		})
-	}
-}
 
 func TestGetCsiDriver(t *testing.T) {
 	ctx := context.Background()
@@ -201,126 +105,3 @@ func TestGetNode(t *testing.T) {
 	}
 }
 
-// makes a csm object with tolerations
-func csmWithTolerations() csmv1.ContainerStorageModule {
-	res := shared.MakeCSM("csm", "driver-test", shared.ConfigVersion)
-
-	// Add tolerations to controller and node
-	res.Spec.Driver.Node.Tolerations = []corev1.Toleration{
-		{
-			Key:               "123",
-			Value:             "123",
-			TolerationSeconds: new(int64),
-		},
-	}
-	res.Spec.Driver.Controller.Tolerations = []corev1.Toleration{
-		{
-			Key:               "123",
-			Value:             "123",
-			TolerationSeconds: new(int64),
-		},
-	}
-
-	// Add FSGroupPolicy
-	res.Spec.Driver.CSIDriverSpec.FSGroupPolicy = "File"
-
-	// Add DNS Policy for GetNode test
-	res.Spec.Driver.DNSPolicy = "ThisIsADNSPolicy"
-
-	// Add image name
-	res.Spec.Driver.Common.Image = "thisIsAnImage"
-
-	// Add pscale driver version
-	res.Spec.Driver.ConfigVersion = "v2.2.0"
-
-	// Add pscale driver version
-	res.Spec.Driver.CSIDriverType = csmv1.PowerScale
-
-	// Add NodeSelector to node and controller
-	res.Spec.Driver.Node.NodeSelector = map[string]string{"thisIs": "NodeSelector"}
-	res.Spec.Driver.Controller.NodeSelector = map[string]string{"thisIs": "NodeSelector"}
-
-	// Add log level to cover some code in GetConfigMap
-	envVarLogLevel := corev1.EnvVar{Name: "CSI_LOG_LEVEL"}
-	res.Spec.Driver.Common.Envs = []corev1.EnvVar{envVarLogLevel}
-	// Add sidecars to trigger code in controller
-	sideCarObjEnabledNil := csmv1.ContainerTemplate{
-		Name:    "driver",
-		Enabled: nil,
-		Args:    []string{"--v=5"},
-	}
-	sideCarObjEnabledFalse := csmv1.ContainerTemplate{
-		Name:    "resizer",
-		Enabled: &falseBool,
-		Args:    []string{"--v=5"},
-	}
-	sideCarObjEnabledTrue := csmv1.ContainerTemplate{
-		Name:    "provisioner",
-		Enabled: &trueBool,
-		Args:    []string{"--volume-name-prefix=k8s"},
-	}
-	sideCarList := []csmv1.ContainerTemplate{sideCarObjEnabledNil, sideCarObjEnabledFalse, sideCarObjEnabledTrue}
-	res.Spec.Driver.SideCars = sideCarList
-
-	return res
-}
-
-// makes a csm object with tolerations
-func csmForPowerScale() csmv1.ContainerStorageModule {
-	res := shared.MakeCSM("csm", "driver-test", shared.ConfigVersion)
-
-	// Add log level to cover some code in GetConfigMap
-	envVarLogLevel1 := corev1.EnvVar{Name: "CERT_SECRET_COUNT", Value: "0"}
-	envVarLogLevel2 := corev1.EnvVar{Name: "X_CSI_ISI_SKIP_CERTIFICATE_VALIDATION", Value: "false"}
-	res.Spec.Driver.Common.Envs = []corev1.EnvVar{envVarLogLevel1, envVarLogLevel2}
-	res.Spec.Driver.AuthSecret = "csm-creds"
-
-	// Add pscale driver version
-	res.Spec.Driver.ConfigVersion = "v2.2.0"
-	res.Spec.Driver.CSIDriverType = csmv1.PowerScale
-
-	return res
-}
-
-// makes a csm object with tolerations
-func csmForPowerScaleBadSkipCert() csmv1.ContainerStorageModule {
-	res := shared.MakeCSM("csm", "driver-test", shared.ConfigVersion)
-
-	// Add log level to cover some code in GetConfigMap
-	envVarLogLevel1 := corev1.EnvVar{Name: "CERT_SECRET_COUNT", Value: "2"}
-	envVarLogLevel2 := corev1.EnvVar{Name: "X_CSI_ISI_SKIP_CERTIFICATE_VALIDATION", Value: "NotABool"}
-	res.Spec.Driver.Common.Envs = []corev1.EnvVar{envVarLogLevel1, envVarLogLevel2}
-
-	// Add pscale driver version
-	res.Spec.Driver.ConfigVersion = "v2.2.0"
-	res.Spec.Driver.CSIDriverType = csmv1.PowerScale
-
-	return res
-}
-
-// makes a csm object with tolerations
-func csmForPowerScaleBadCertCnt() csmv1.ContainerStorageModule {
-	res := shared.MakeCSM("csm", "driver-test", shared.ConfigVersion)
-
-	// Add log level to cover some code in GetConfigMap
-	envVarLogLevel1 := corev1.EnvVar{Name: "CERT_SECRET_COUNT", Value: "thisIsNotANumber"}
-	envVarLogLevel2 := corev1.EnvVar{Name: "X_CSI_ISI_SKIP_CERTIFICATE_VALIDATION", Value: "true"}
-	res.Spec.Driver.Common.Envs = []corev1.EnvVar{envVarLogLevel1, envVarLogLevel2}
-
-	// Add pscale driver version
-	res.Spec.Driver.ConfigVersion = "v2.2.0"
-	res.Spec.Driver.CSIDriverType = csmv1.PowerScale
-
-	return res
-}
-
-// makes a csm object with tolerations
-func csmForPowerScaleBadVersion() csmv1.ContainerStorageModule {
-	res := shared.MakeCSM("csm", "driver-test", shared.ConfigVersion)
-
-	// Add pscale driver version
-	res.Spec.Driver.ConfigVersion = "v0"
-	res.Spec.Driver.CSIDriverType = csmv1.PowerScale
-
-	return res
-}
